executor: nitpick for implementing meldata for rank funcs to track memUsage
